+++ coro_rpc/server/rpc_server.h
@@ -87,7 +87,7 @@ struct Method {
   // (boost::asio::awaitable) so it may co_await work before calling exactly one
   // of reply/error_reply.
   Method(RpcServer *server, std::string name, std::string request_type,
-         std::string response_type, int32_t slot_size, int32_t num_slots,
+         std::string response_type, int64_t slot_size, int32_t num_slots,
          std::function<boost::asio::awaitable<void>(
              const google::protobuf::Any &,
              std::function<void(std::unique_ptr<google::protobuf::Any>)>,
@@ -101,7 +101,7 @@ struct Method {
   }
 
   Method(RpcServer *server, std::string name, std::string request_type,
-         std::string response_type, int32_t slot_size, int32_t num_slots,
+         std::string response_type, int64_t slot_size, int32_t num_slots,
          std::function<boost::asio::awaitable<absl::Status>(
              const google::protobuf::Any &, internal::AnyStreamWriter &)>
              callback,
@@ -119,7 +119,7 @@ struct Method {
   std::string name;
   std::string request_type;
   std::string response_type;
-  int32_t slot_size;
+  int64_t slot_size;
   int32_t num_slots;
   // Async handler for a normal, non-streaming method.  See the constructor
   // above for the calling convention.
